ctai-ctd-hacks/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pymongo==4.8.0
python-dotenv==1.0.1
beautifulsoup4==4.12.0
requests==2.32.0
fake-useragent==1.5.0
pandas==2.1.4
scikit-learn==1.4.2
pydantic==2.8.0